refactor(main-api): 패키지 구조 변경
- main-api > query > member 도메인 패키지 이동
