+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/member/application/query/MembersInRoom.java
@@ -1,4 +1,4 @@
-package com.whoz_in.main_api.query.device.application.active;
+package com.whoz_in.main_api.query.member.application.query;
 
 import com.whoz_in.main_api.query.shared.application.Query;
