+++ modules/main-api/src/main/java/com/whoz_in/main_api/query/member/application/response/MembersInRoomResponse.java
@@ -1,4 +1,4 @@
-package com.whoz_in.main_api.query.device.application.active;
+package com.whoz_in.main_api.query.member.application.response;
 
 import com.whoz_in.main_api.query.shared.application.Response;
 import java.util.List;
